day8/infrawatch-auth/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
sqlalchemy==2.0.16
alembic==1.11.1
bcrypt==4.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
marshmallow==3.19.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
uuid==1.30
pytest==7.4.0
pytest-cov==4.1.0
factory-boy==3.2.1
flask-testing==0.8.1
requests==2.31.0
gunicorn==20.1.0
